pls work

+++ base_firmware/src/central.h
@@ -7,7 +7,7 @@
 #define NUS_MAX_DATA_LEN 13
 // central.h
 struct bt_data_recieved {
-    uint16_t data_len;
+    size_t data_len;
     int8_t data_buffer[NUS_MAX_DATA_LEN]; // e.g. #define NUS_MAX_DATA_LEN 244
 };
